style: restyling n restructuring for clean code
