+++ Otterator/View/TranscriptView.swift
@@ -17,48 +17,24 @@ struct TranscriptView: View {
     var tabs = ["Transcript","Summary"]
     var body: some View {
         NavigationStack{
-            Picker("segment",selection:$segment){
-                ForEach(tabs, id:\.self){
-                    Text($0)
-                }
-            }
-            .pickerStyle(.segmented)
-            .padding(16)
-            CustomTranscript(progress:$progress)
-            .padding(.horizontal,20)
-            .padding(.vertical,8)
-            Spacer()
             VStack{
-                CustomProgressView(progress:$progress)
-                    .frame(height:20)
-                    .padding(.horizontal,16)
-                HStack(alignment:.center){
-                    Text(formatSec(progress * duration))
-                    Spacer()
-                    Text(formatSec(duration))
-                }.padding(.horizontal,6)
-            }
-            .padding(.horizontal,16)
-            .padding(.vertical,12)
-            HStack{
-                Spacer()
-                Image(systemName:"gobackward.15")
-                    .resizable()
-                    .frame(width:40,height:40)
-                Spacer()
-                Image(systemName: !isPlay ? "play.fill" : "pause.fill")
-                    .resizable()
-                    .frame(width:40,height:40)
-                    .onTapGesture{
-                        withAnimation(.easeIn){
-                            isPlay = !isPlay
-                        }
+                Picker("segment",selection:$segment){
+                    ForEach(tabs, id:\.self){
+                        Text($0)
                     }
-                Spacer()
-                Image(systemName:"goforward.15")
-                    .resizable()
-                    .frame(width:40,height:40)
-                Spacer()
+                }
+                .pickerStyle(.segmented)
+                .padding(16)
+                
+                if segment == "Transcript" {
+                    TranscriptContentView()
+                } else {
+                    VStack{
+                        Spacer()
+                        Text("Summary Page")
+                        Spacer()
+                    }
+                }
             }
             .navigationTitle("Recording 001")
             .navigationBarTitleDisplayMode(.inline)
@@ -88,19 +64,12 @@ struct TranscriptView: View {
                 }
             }
             .sheet(isPresented:$isPresented[2]){
-                    InformationModalView()
-                        .presentationDetents([.large])
-                        .presentationBackgroundInteraction(.disabled)
-                        .presentationBackground(.ultraThickMaterial)
+                InformationModalView()
+                    .presentationDetents([.large])
+                    .presentationBackgroundInteraction(.disabled)
+                    .presentationBackground(.ultraThickMaterial)
             }
         }
-    }
-    func formatSec(_ second:Double) -> String
-    {
-        let totalSeconds = Int(floor(second))
-        let minutes = totalSeconds / 60
-        let remainingSeconds = totalSeconds % 60
-        return String(format: "%02d:%02d", minutes, remainingSeconds)
     }
 }
